refactor: redux ducks 패턴에서 actions, constants, reducers로 디렉토리 분리
